fix: fixed deprecations and ran `clippy --fix`

+++ aw-models/src/event.rs
@@ -41,7 +41,7 @@ impl Event {
     }
     pub fn calculate_endtime(&self) -> DateTime<Utc> {
         self.timestamp
-            + chrono::Duration::nanoseconds(self.duration.num_nanoseconds().unwrap() as i64)
+            + chrono::Duration::nanoseconds(self.duration.num_nanoseconds().unwrap())
     }
     pub fn interval(&self) -> TimeInterval {
         TimeInterval::new(self.timestamp, self.calculate_endtime())
